+++ HISTORY.md
@@ -48,6 +48,7 @@ Hypotenous / |
 - [x] Draw Filled Triangle
 - [x] Add controls to manage culling and rendering modes
 - [x] Add Face (Vec3 + Color) to represent pre-projected triangle
+- [x] Add Painters algorithm
 
 Normals are for lighting, but also for backface culling.